requirements.txt
langchain-groq
langchain
langchain_community
streamlit
python-dotenv
PyPDF2

-e .